refactor: field and field type
