--- lib/wraft_doc/documents/field_type_enum.ex
@@ -1,2 +1,0 @@
-import EctoEnum
-defenum(WraftDoc.Documents.FieldTypeEnum, ["string", "number", "date"])
