+++ lib/wraft_doc/fields/field_type_enum.ex
@@ -1,0 +1,2 @@
+import EctoEnum
+defenum(WraftDoc.Fields.FieldTypeEnum, ["string", "number", "date"])
